+++ scripts/test/DynamicField/ValueValidate.t
@@ -384,28 +384,64 @@ my @Tests = (
         Success => 0,
     },
     {
-        Name   => 'Long Value Dropdown',
+        Name   => 'Wrong Value Dropdown',
         Config => {
             DynamicFieldConfig => $DynamicFieldConfigs{Dropdown},
-            Value              => 'Looooooooooooooooooooooooooooong',
+            Value              => 'Value 1',
+            UserID             => $UserID,
+        },
+        Success => 0,
+    },
+    {
+        Name   => 'Value Dropdown',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldConfigs{Dropdown},
+            Value              => '1',
             UserID             => $UserID,
         },
         Success => 1,
+    },
+    {
+        Name   => 'Wrong Single Value Multiselect',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldConfigs{Multiselect},
+            Value              => 'Value 1',
+            UserID             => $UserID,
+        },
+        Success => 0,
     },
     {
         Name   => 'Single Value Multiselect',
         Config => {
             DynamicFieldConfig => $DynamicFieldConfigs{Multiselect},
-            Value              => 'Value1',
+            Value              => ['1'],
             UserID             => $UserID,
         },
         Success => 1,
     },
     {
+        Name   => 'Wrong Single Value Multiselect',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldConfigs{Multiselect},
+            Value              => [ '1', 'Value 2' ],
+            UserID             => $UserID,
+        },
+        Success => 0,
+    },
+    {
+        Name   => 'Wrong Multiple Values Multiselect',
+        Config => {
+            DynamicFieldConfig => $DynamicFieldConfigs{Multiselect},
+            Value              => [ 'Value 1', 'Value 2' ],
+            UserID             => $UserID,
+        },
+        Success => 0,
+    },
+    {
         Name   => 'Multiple Values Multiselect',
         Config => {
             DynamicFieldConfig => $DynamicFieldConfigs{Multiselect},
-            Value              => [ 'Value1', 'Value2' ],
+            Value              => [ '1', '2' ],
             UserID             => $UserID,
         },
         Success => 1,